machine: grinn-genio-700-sbc: remove unused machine overrides
`mt8188-evk` was not actually used and did not contribute
to any build and image changes.

`genio-700-evk` provided 3 board-specific recipes:
 - dtbo - no longer required, as dedicated DTBOs are now provided
 - alsa-state - no longer required, as dedicated
configuration is now provided
 - base-files - board-specific files for USB ETH gadget configuration
are added, but does not seem important for `Grinn GenioSBC-700`.

Only left additional machine override `mt8390-evk` is kept,
as it provides Linux kernel configuration, which is not
yet provided for the board.