bsp: linux: evb: configure RGMII interface timing
The RGMII interface requires specific clock-to-data delays for reliable
operation. Since the PCB does not introduce any trace delays and the
PHY's maximum configurable skew is insufficient to meet RGMII timing
requirements, the TX delay is instead provided by the MAC, as specified
in the PHY’s datasheet. This is supported by the MediaTek DWMAC wrapper,
which allows internal delay configuration.

The default RX delay introduced by the PHY is 1.2 ns. An additional delay
of 0.6 ns has been configured, resulting in a total RX delay of 1.8 ns.

On the TX path, the default delay introduced by the PHY is 0 ns. A total
delay of 1.74 ns has been configured through the MAC’s internal skew
settings.

All skew values are now explicitly defined to ensure the implemented
timing is unambiguous and fully under control.