fixes: trusted-firmware-a: fix fiptool do compile
To properly build the fiptool, which is a build host side tool,
path to the include and lib directories have to be provided.
Without passing the BUILD_* variables fiptool won't build.

MTK uses trusted-firmware-a.inc from meta-arm that is designed for
a newer version of tfa (2.10). MTK uses an older version of tfa
(2.6). In the newer realse of tfa the LDLIBS variable was replaced
with LDOPTS, hence the fix does not apply. Therefore append
BUILD_LDFLAGS to the LDLIBS variable.