bsp: linux: fix and restructure
cp in the do_configure breaks devtool build. Files are not in the
same place as during the normal bitbake usage. Moreover, cp in
do_configure doesnt seem to be used by anyone.

Therefore to fix this the "subdir=" option of SRC_URI was used to
directly unpack the dts files into the linux kernel source code.
This not only fixes the problem with devtool but also
significantly simplifies the bbappend file.

The fix is based on the following issue:
https://community.nxp.com/t5/MCX-Microcontrollers/Devtool-modify-failed/td-p/1837201